Extract DeviceGrant implementation into module mixin
An app wishing to customize the DeviceGrant model class can now reuse
the default implementation by including a module rather than extending
from the base class. This allows the model class to use the [multiple
database][1] support in Rails 6.1, which requires extending from an
abstract class descending from ActiveRecord::Base.

[1]: https://guides.rubyonrails.org/active_record_multiple_databases.html

```
class AnimalsRecord < ActiveRecord::Base
  self.abstract_class = true
  connects_to database: { writing: :animals, reading: :animals }
end

class DeviceGrant < AnimalsRecord
  include Doorkeeper::DeviceAuthorizationGrant::DeviceGrantMixin
  self.table_name = "oauth_device_grant"
end

Doorkeeper::DeviceAuthorizationGrant.configure do
  device_grant_class "DeviceGrant"
end
```

This structure is borrowed from [doorkeeper][2] which also uses mixins.

[2]: https://github.com/doorkeeper-gem/doorkeeper/tree/4c1a5719d3f6f0ef7d96c2531b66b8c2e2c30dc8/lib/doorkeeper/models
